NH-3272 - Move TypeHelperExtensionMethods to a more internal namespace (Part 2)
