+++ src/NHibernate/Linq/Visitors/HqlGeneratorExpressionTreeVisitor.cs
@@ -5,8 +5,9 @@ using NHibernate.Engine.Query;
 using NHibernate.Hql.Ast;
 using NHibernate.Linq.Expressions;
 using NHibernate.Linq.Functions;
-using NHibernate.Linq.Internal;
 using NHibernate.Param;
+using NHibernate.Util;
+
 using Remotion.Linq.Clauses.Expressions;
 using Remotion.Linq.Clauses.ResultOperators;
