+++ src/NHibernate/NHibernate.csproj
@@ -715,6 +715,7 @@
     <Compile Include="Util\SequencedHashMap.cs" />
     <Compile Include="Util\StringHelper.cs" />
     <Compile Include="Util\StringTokenizer.cs" />
+    <Compile Include="Util\TypeExtensions.cs" />
     <Compile Include="WrongClassException.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -951,7 +952,6 @@
     <Compile Include="Linq\Functions\LinqToHqlGeneratorsRegistryFactory.cs" />
     <Compile Include="Linq\Functions\StandardLinqExtensionMethodGenerator.cs" />
     <Compile Include="Linq\LinqExtensionMethodAttribute.cs" />
-    <Compile Include="Linq\TypeHelperExtensionMethods.cs" />
     <Compile Include="Linq\Visitors\NameUnNamedParameters.cs" />
     <Compile Include="Linq\NhRelinqQueryParser.cs" />
     <Compile Include="Linq\Expressions\NhAggregatedExpression.cs" />
