+++ src/NHibernate/Util/TypeExtensions.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Collections.Generic;
+
+namespace NHibernate.Util
+{
+	public static class TypeExtensions
+	{
+		public static bool IsEnumerableOfT(this System.Type type)
+		{
+			return type.IsGenericType && type.GetGenericTypeDefinition() == typeof(IEnumerable<>);
+		}
+
+		public static bool IsNullable(this System.Type type)
+		{
+			return (type.IsGenericType && type.GetGenericTypeDefinition() == typeof(Nullable<>));
+		}
+
+		public static bool IsNullableOrReference(this System.Type type)
+		{
+			return !type.IsValueType || type.IsNullable();
+		}
+
+		public static System.Type NullableOf(this System.Type type)
+		{
+			return type.GetGenericArguments()[0];
+		}
+
+		public static bool IsPrimitive(this System.Type type)
+		{
+			return (type.IsValueType || type.IsNullable() || type == typeof(string));
+		}
+
+		public static bool IsNonPrimitive(this System.Type type)
+		{
+			return !type.IsPrimitive();
+		}
+	}
+}
